requirements.txt
bottle==0.13.4
bottle-file-cache==1.0.0
cryptography==45.0.5
jinja2==3.1.6
python-ulid==3.0.0
requests==2.32.4
secure==1.0.1
sentry-sdk==2.32.0
